file structure for entry point, parent and children
