+++ src/spas/home/components/ScoreCard/index.js
@@ -1,8 +1,18 @@
 import React, { Component } from 'react';
+import '../../assets/home.font';
 
 class ScoreCard extends Component {
+  setWinningScore() {
+    let tally = [];
+    for (let i = 0; i < 10; i++) {
+      tally.push(<span key={`tally-${i}`} className="icon icon-beer" />);
+    }
+
+    return tally;
+  }
+
   render() {
-    return <div>{this.props.children}</div>;
+    return <div> {this.setWinningScore()} </div>;
   }
 }
